# contracts/api-contract-draft.yaml
openapi: 3.1.0
info:
  title: Patika Beosztás API
  version: 0.3.0-phase2b
  description: >-
    A Phase 2B frontend–backend szerződése. A futó backend
    /openapi/v1.json dokumentuma a generált frontend típusok végleges forrása.
    A frontend HTTPS-en, az API-val azonos site alatt fusson, és a cookie-s
    kérésekhez credentials: include beállítást használjon.
servers:
  - url: /
...
paths:
...
  /api/admin/employees:
    get:
      tags: [Employees]
      summary: Dolgozók listázása
      x-required-permission: ManageEmployees
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/IncludeInactive'
      responses:
        '200':
          description: Lapozott dolgozólista
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedEmployees'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
    post:
      tags: [Employees]
      summary: Dolgozó létrehozása
      x-required-permission: ManageEmployees
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Létrehozott dolgozó
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /api/admin/employees/{id}:
    parameters:
      - $ref: '#/components/parameters/EntityId'
    get:
      tags: [Employees]
      summary: Dolgozó lekérése
      x-required-permission: ManageEmployees
      responses:
        '200':
          description: Dolgozó
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '404':
          $ref: '#/components/responses/Problem'
    put:
      tags: [Employees]
      summary: Dolgozó módosítása vagy deaktiválása
      x-required-permission: ManageEmployees
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
      responses:
        '200':
          description: Módosított dolgozó
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
...
  /api/admin/employees/{employeeId}/shift-quota-rules:
    parameters:
      - $ref: '#/components/parameters/EmployeeId'
    get:
      tags: [Employee shift quota rules]
      summary: Dolgozó műszakkvóta-szabályainak listázása
      x-required-permission: ManageEmployees
      parameters:
        - $ref: '#/components/parameters/IncludeInactive'
      responses:
        '200':
          description: Műszakkvóta-szabályok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeShiftQuotaRuleResponse'
        '404': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Employee shift quota rules]
      summary: Dolgozói műszakkvóta-szabály létrehozása
      x-required-permission: ManageEmployees
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeShiftQuotaRuleRequest'
      responses:
        '201':
          description: Létrehozott műszakkvóta-szabály
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeShiftQuotaRuleResponse'
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/admin/employee-shift-quota-rules/{id}:
    parameters:
      - $ref: '#/components/parameters/EntityId'
    put:
      tags: [Employee shift quota rules]
      summary: Dolgozói műszakkvóta-szabály módosítása
      x-required-permission: ManageEmployees
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeShiftQuotaRuleRequest'
      responses:
        '200':
          description: Módosított műszakkvóta-szabály
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeShiftQuotaRuleResponse'
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/admin/employee-shift-quota-rules/{id}/deactivate:
    parameters:
      - $ref: '#/components/parameters/EntityId'
    post:
      tags: [Employee shift quota rules]
      summary: Dolgozói műszakkvóta-szabály inaktiválása
      x-required-permission: ManageEmployees
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionRequest'
      responses:
        '200':
          description: Inaktivált műszakkvóta-szabály
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeShiftQuotaRuleResponse'
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-PatikaSession
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-TOKEN
      required: true
      schema:
        type: string
    EntityId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    EmployeeId:
      in: path
      name: employeeId
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Search:
      in: query
      name: search
      schema:
        type: [string, 'null']
    IncludeInactive:
      in: query
      name: includeInactive
      schema:
        type: boolean
        default: false
  responses:
    Problem:
      description: RFC 7807 ProblemDetails
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    ProfessionalRole:
      type: string
      enum:
        - PharmacyManager
        - Pharmacist
        - SpecialistAssistant
        - Assistant
        - PharmacistTrainee
        - AssistantTrainee
        - Cleaner
        - FinanceHelper
        - Other
...
    EmployeeTimeWindowType:
      type: string
      enum: [Preferred, Forbidden]
    TimeType:
      type: string
      enum:
        - Work
        - Overtime
        - OnCallDuty
        - Standby
        - AnnualLeave
        - SickLeave
        - UnpaidLeave
        - ParentalLeave
        - Other
    ProblemDetails:
      type: object
      properties:
        type:
          type: [string, 'null']
        title:
          type: [string, 'null']
        status:
          type: [integer, 'null']
        detail:
          type: [string, 'null']
        instance:
          type: [string, 'null']
        code:
          type: [string, 'null']
        errors:
          type: [array, 'null']
          items:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
              field: { type: [string, 'null'] }
...
    EmployeeLocationInput:
      type: object
      required: [locationId]
      properties:
        locationId: { type: string, format: uuid }
        enabled: { type: boolean, default: true }
    EmployeeTimeWindowInput:
      type: object
      required: [dayOfWeek, startTime, endTime, type]
      properties:
        dayOfWeek:
          type: [string, 'null']
          enum: [Sunday, Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, null]
        startTime: { type: string, format: time }
        endTime: { type: string, format: time }
        type:
          $ref: '#/components/schemas/EmployeeTimeWindowType'
    CreateEmployeeRequest:
      type: object
      required:
        - fullName
        - displayName
        - professionalRole
        - isActive
        - isSchedulable
        - includeInAutoFill
        - countsAsPharmacist
      properties:
        fullName: { type: string, minLength: 1, maxLength: 200 }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        professionalRole:
          $ref: '#/components/schemas/ProfessionalRole'
        isActive: { type: boolean }
        isSchedulable: { type: boolean }
        includeInAutoFill: { type: boolean }
        countsAsPharmacist: { type: boolean }
        monthlyMinutesLimit: { type: [integer, 'null'], minimum: 1, maximum: 44640 }
        maxDailyMinutes: { type: [integer, 'null'], minimum: 1, maximum: 1440 }
        birthDate:
          type: [string, 'null']
          format: date
          description: Nem lehet jövőbeli; a legkorábbi elfogadott dátum 1900-01-01.
        externalPayrollId: { type: [string, 'null'], maxLength: 100 }
        locations:
          type: [array, 'null']
          items:
            $ref: '#/components/schemas/EmployeeLocationInput'
        timeWindows:
          type: [array, 'null']
          items:
            $ref: '#/components/schemas/EmployeeTimeWindowInput'
        allowedTimeTypes:
          type: [array, 'null']
          uniqueItems: true
          items:
            $ref: '#/components/schemas/TimeType'
    UpdateEmployeeRequest:
      allOf:
        - $ref: '#/components/schemas/CreateEmployeeRequest'
        - type: object
          required: [expectedVersion]
          properties:
            expectedVersion:
              type: integer
              minimum: 1
    EmployeeLocationResponse:
      type: object
      required: [locationId, locationName, enabled]
      properties:
        locationId: { type: string, format: uuid }
        locationName: { type: string }
        enabled: { type: boolean }
    EmployeeTimeWindowResponse:
      allOf:
        - $ref: '#/components/schemas/EmployeeTimeWindowInput'
        - type: object
          required: [id]
          properties:
            id: { type: string, format: uuid }
    LinkedUserSummary:
      type: object
      required: [userId, email, displayName, isActive]
      properties:
        userId: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        isActive: { type: boolean }
    EmployeeResponse:
      type: object
      required:
        - id
        - fullName
        - displayName
        - professionalRole
        - isActive
        - isSchedulable
        - includeInAutoFill
        - countsAsPharmacist
        - locations
        - timeWindows
        - allowedTimeTypes
        - linkedUser
        - warnings
        - version
        - createdAtUtc
        - updatedAtUtc
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
        displayName: { type: string }
        professionalRole:
          $ref: '#/components/schemas/ProfessionalRole'
        isActive: { type: boolean }
        isSchedulable: { type: boolean }
        includeInAutoFill: { type: boolean }
        countsAsPharmacist: { type: boolean }
        monthlyMinutesLimit: { type: [integer, 'null'] }
        maxDailyMinutes: { type: [integer, 'null'] }
        birthDate: { type: [string, 'null'], format: date }
        externalPayrollId: { type: [string, 'null'] }
        locations:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeLocationResponse'
        timeWindows:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeTimeWindowResponse'
        allowedTimeTypes:
          type: array
          items:
            $ref: '#/components/schemas/TimeType'
        linkedUser:
          oneOf:
            - $ref: '#/components/schemas/LinkedUserSummary'
            - type: 'null'
        warnings:
          type: array
          items: { type: string }
        version: { type: integer, minimum: 1 }
        createdAtUtc: { type: string, format: date-time }
        updatedAtUtc: { type: string, format: date-time }
...
    VersionRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, minimum: 1 }
...
    ShiftQuotaDimension:
      type: string
      enum: [MorningShift, AfternoonShift, EveningShift, LongShift, SaturdayShift, SundayShift, OnCallDuty, Standby]
    QuotaPeriod:
      type: string
      enum: [Week, Month]
    QuotaSeverity:
      type: string
      enum: [Preferred, Required]
...
    CreateEmployeeShiftQuotaRuleRequest:
      type: object
      additionalProperties: false
      required: [dimension, period, minimum, target, maximum, severity, isActive]
      properties:
        dimension: { $ref: '#/components/schemas/ShiftQuotaDimension' }
        period: { $ref: '#/components/schemas/QuotaPeriod' }
        minimum: { type: integer, minimum: 0 }
        target: { type: integer, minimum: 0 }
        maximum: { type: integer, minimum: 0 }
        severity: { $ref: '#/components/schemas/QuotaSeverity' }
        isActive: { type: boolean }
    UpdateEmployeeShiftQuotaRuleRequest:
      allOf:
        - $ref: '#/components/schemas/CreateEmployeeShiftQuotaRuleRequest'
        - type: object
          required: [expectedVersion]
          properties:
            expectedVersion: { type: integer, minimum: 1 }
    EmployeeShiftQuotaRuleResponse:
      allOf:
        - $ref: '#/components/schemas/CreateEmployeeShiftQuotaRuleRequest'
        - type: object
          required: [id, employeeId, employeeDisplayName, version, createdAtUtc, updatedAtUtc]
          properties:
            id: { type: string, format: uuid }
            employeeId: { type: string, format: uuid }
            employeeDisplayName: { type: string }
            version: { type: integer, minimum: 1 }
            createdAtUtc: { type: string, format: date-time }
            updatedAtUtc: { type: string, format: date-time }
...
    PagedEmployees:
      type: object
      required: [items, page, pageSize, totalCount]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeResponse'
        page: { type: integer }
        pageSize: { type: integer }
        totalCount: { type: integer }
...
security:
  - cookieAuth: []